Shove all resizing responsibility on the calling scope
I need to maintain percentage widths, but with two sources of truth
(the state of the Resizeable vs. the state of the calling scope), a race
condition made it ambiguous if the final style state would be % or px.

Didn't change x and y because I only care about dimensions for now, but
presumably the same changes would apply there.